Keywords are now unique

+++ include/mocha/list.h
@@ -10,7 +10,7 @@ typedef struct mocha_list {
 	size_t count;
 } mocha_list;
 
-void mocha_list_init(mocha_list* self, const struct mocha_object* args[], int count);
+void mocha_list_init(mocha_list* self, const struct mocha_object* args[], size_t count);
 mocha_boolean mocha_list_equal(const mocha_list* self, const mocha_list* other);
 
 #endif
