+++ include/mocha/vector.h
@@ -10,7 +10,7 @@ typedef struct mocha_vector {
 	size_t count;
 } mocha_vector;
 
-void mocha_vector_init(mocha_vector* self, const struct mocha_object* args[], int count);
+void mocha_vector_init(mocha_vector* self, const struct mocha_object* args[], size_t count);
 mocha_boolean mocha_vector_equal(const mocha_vector* self, const mocha_vector* other);
 
 #endif
