+++ lib/vector.c
@@ -2,7 +2,7 @@
 #include <mocha/object.h>
 #include <stdlib.h>
 
-void mocha_vector_init(mocha_vector* self, const struct mocha_object* args[], int count)
+void mocha_vector_init(mocha_vector* self, const struct mocha_object* args[], size_t count)
 {
 	size_t octet_count = sizeof(mocha_object*) * count;
 	self->objects = malloc(octet_count);
@@ -16,7 +16,7 @@ mocha_boolean mocha_vector_equal(const mocha_vector* self, const mocha_vector* o
 		return mocha_false;
 	}
 
-	for (size_t i=0; i<self->count; ++i) {
+	for (size_t i = 0; i < self->count; ++i) {
 		if (!mocha_object_equal(self->objects[i], other->objects[i])) {
 			return mocha_false;
 		}
